Add files via upload
Description
The code consists of a quiz about the animated series "Rick and Morty". The quiz is implemented using HTML, CSS, and JavaScript. JavaScript is responsible for calculating the user's score based on the provided answers and displaying the result.

Features
Answer Selection: Users can select an answer option for each question in the quiz.
Score Calculation: The code calculates the user's score based on the correct answers provided.
Result Display: The quiz result is shown to the user, indicating the percentage of correct answers.
Prerequisites
Before running the code, make sure you have a compatible web browser installed on your device.

How to Run
Follow the steps below to run the quiz code:

Download the HTML, CSS, and JavaScript files of the quiz.
Open the HTML file (index.html) in a compatible web browser.
Usage Example
When running the quiz code, you will be presented with a series of questions about Rick and Morty. Select the correct option for each question and click the "Submit" button. The result will be displayed, showing the percentage of correct answers.